Break specialized graphics into specific mods
